Move implementation to cpp file.
